refactor(unlock-app): Remove useAuthenticateHandler and replace with useAuthenticate method
