+++ unlock-app/src/hooks/useAuthenticate.ts
@@ -1,8 +1,8 @@
 import WalletConnectProvider from '@walletconnect/web3-provider'
-import { useContext } from 'react'
 import WalletLink from 'walletlink'
-import { ConfigContext } from '../utils/withConfig'
-import { AuthenticationContext } from '../contexts/AuthenticationContext'
+import { useConfig } from '../utils/withConfig'
+import { useAuth } from '../contexts/AuthenticationContext'
+import { useAppStorage } from './useAppStorage'
 
 export interface EthereumWindow extends Window {
   ethereum?: any
@@ -46,9 +46,20 @@ interface AuthenticateProps {
   injectedProvider?: any | null
 }
 
-export function useAuthenticate({ injectedProvider }: AuthenticateProps) {
-  const config = useContext(ConfigContext)
-  const { authenticate } = useContext(AuthenticationContext)
+export enum WALLET_PROVIDER {
+  METAMASK,
+  WALLET_CONNECT,
+  COINBASE,
+  UNLOCK,
+}
+
+export type WalletProvider = keyof typeof WALLET_PROVIDER
+
+export function useAuthenticate(options: AuthenticateProps = {}) {
+  const { injectedProvider } = options
+  const config = useConfig()
+  const { authenticate } = useAuth()
+  const { setStorage, removeKey } = useAppStorage()
 
   const injectedOrDefaultProvider = injectedProvider || selectProvider(config)
 
@@ -77,6 +88,38 @@ export function useAuthenticate({ injectedProvider }: AuthenticateProps) {
     return authenticate(ethereum)
   }
 
+  const walletHandlers: {
+    [key in WalletProvider]: (provider?: any) => Promise<any | void>
+  } = {
+    METAMASK: handleInjectProvider,
+    WALLET_CONNECT: handleWalletConnectProvider,
+    COINBASE: handleCoinbaseWalletProvider,
+    UNLOCK: handleUnlockProvider,
+  }
+
+  async function authenticateWithProvider(
+    providerType: WalletProvider,
+    provider?: any
+  ) {
+    if (!walletHandlers[providerType]) {
+      removeKey('provider')
+    }
+    const connectedProvider = walletHandlers[providerType](provider)
+
+    connectedProvider.then((p) => {
+      if (!p?.account) {
+        return console.error('Unable to get provider')
+      }
+      if (p?.isUnlock && p?.email) {
+        setStorage('email', p.email)
+      } else {
+        removeKey('email')
+      }
+      setStorage('provider', providerType)
+    })
+    return connectedProvider
+  }
+
   return {
     handleUnlockProvider,
     handleInjectProvider,
@@ -84,5 +127,6 @@ export function useAuthenticate({ injectedProvider }: AuthenticateProps) {
     handleCoinbaseWalletProvider,
     injectedOrDefaultProvider,
     authenticate,
+    authenticateWithProvider,
   }
 }
